+++ Assets/NativeStringCollections/Scripts/AsyncTextFileLoader.cs
@@ -50,8 +50,17 @@ namespace NativeStringCollections
     public readonly struct ReadState
     {
         public ReadJobState JobState { get; }
+        /// <summary>
+        /// total length by BrockSize.
+        /// </summary>
         public int Length { get; }
+        /// <summary>
+        /// progress by BrockSize.
+        /// </summary>
         public int Read { get; }
+        /// <summary>
+        /// total reference count.
+        /// </summary>
         public int RefCount { get; }
         /// <summary>
         /// elapsed milliseconds for AsyncReadManager.Read()
@@ -70,7 +79,13 @@ namespace NativeStringCollections
         /// elapsed milliseconds to parse the file.
         /// </summary>
         public double Delay { get { return DelayReadAsync + DelayParseText + DelayPostProc; } }
+        /// <summary>
+        /// loading data is completed or not.
+        /// </summary>
         public bool IsCompleted { get { return (JobState == ReadJobState.Completed); } }
+        /// <summary>
+        /// the JobState is Completed or UnLoaded. it must be true to access parser class.
+        /// </summary>
         public bool IsStandby
         {
             get { return (JobState == ReadJobState.Completed || JobState == ReadJobState.UnLoaded); }
@@ -158,7 +173,7 @@ namespace NativeStringCollections
                 _parser.Dispose();
                 _state.Dispose();
 
-                _allocated = true;
+                _allocated = false;
             }
         }
 
@@ -172,10 +187,24 @@ namespace NativeStringCollections
         }
 
         public unsafe ReadState GetState { get { return _state.Target->GetState(); } }
+        public bool IsCompleted
+        {
+            get { return GetState.JobState == ReadJobState.Completed; }
+        }
+        public bool IsStandby
+        {
+            get
+            {
+                var job_stat = GetState.JobState;
+                return (job_stat == ReadJobState.Completed || job_stat == ReadJobState.UnLoaded);
+            }
+        }
 
         public JobHandle LoadFile() { return this.LoadFile(Path); }
         public unsafe JobHandle LoadFile(string path)
         {
+            Path = path;
+
             if (path.Length == 0)
                 throw new ArgumentException("path string is empty.");
 
@@ -189,13 +218,17 @@ namespace NativeStringCollections
 
             return _job_handle;
         }
+        public unsafe void Update()
+        {
+            if (_state.Target->JobState == ReadJobState.WaitForCallingComplete)
+            {
+                Complete();
+            }
+        }
         public unsafe void Complete()
         {
-            if(_state.Target->JobState == ReadJobState.WaitForCallingComplete)
-            {
-                _job_handle.Complete();
-                _state.Target->JobState = ReadJobState.Completed;
-            }
+            _job_handle.Complete();
+            _state.Target->JobState = ReadJobState.Completed;
         }
         public unsafe void UnLoadFile()
         {
@@ -209,6 +242,8 @@ namespace NativeStringCollections
         public void LoadFileInMainThread() { this.LoadFileInMainThread(Path); }
         public unsafe void LoadFileInMainThread(string path)
         {
+            Path = path;
+
             if (path.Length == 0)
                 throw new ArgumentException("path string is empty.");
 
@@ -452,6 +487,15 @@ namespace NativeStringCollections
         {
             return _state[index].Target->GetState();
         }
+        public bool IsCompleted(int index)
+        {
+            return GetState(index).JobState == ReadJobState.Completed;
+        }
+        public bool IsStandby(int index)
+        {
+            var stat = GetState(index);
+            return (stat.JobState == ReadJobState.Completed || stat.JobState == ReadJobState.UnLoaded);
+        }
 
         public void LoadFile(int index)
         {
@@ -467,6 +511,55 @@ namespace NativeStringCollections
         {
             this.UpdateImpl(this.FlushLoadJobs);
             this.FlushLoadJobs = false;
+        }
+        /// <summary>
+        /// complete all running jobs.
+        /// </summary>
+        public unsafe void Complete()
+        {
+            for (int i = _runningJob.Length - 1; i >= 0; i--)
+            {
+                var job_info = _runningJob[i];
+
+                _parserPool[job_info.ParserID].Complete();
+                _state[job_info.FileIndex].Target->JobState = ReadJobState.Completed;
+
+                this.ReleaseParser(job_info.ParserID);
+                _runningJob.RemoveAt(i);
+            }
+
+            if (!_unLoadJob.IsCompleted)
+            {
+                _unLoadJob.Complete();
+                _unLoadJob.Clear();
+            }
+        }
+        /// <summary>
+        /// complete if the job running for file[index].
+        /// </summary>
+        /// <param name="index"></param>
+        public unsafe void Complete(int index)
+        {
+            for(int i=0; i<_runningJob.Length; i++)
+            {
+                var job_info = _runningJob[i];
+                if (job_info.FileIndex == index)
+                {
+                    _parserPool[job_info.ParserID].Complete();
+                    _state[index].Target->JobState = ReadJobState.Completed;
+
+                    this.ReleaseParser(job_info.ParserID);
+                    _runningJob.RemoveAt(i);
+
+                    return;
+                }
+            }
+
+            if (!_unLoadJob.IsCompleted)
+            {
+                _unLoadJob.Complete();
+                _unLoadJob.Clear();
+            }
         }